Add frontend for match page data
This commit adds the frontend for the match page data, including the
display of teams and scores with details and information cards. There
is also a timeline that creates new timeline events using an array.
This allows users to easily view and understand the data for a
particular match, providing them with valuable insights and
information. Frontend also shows errors when match is not in database.