Update fixture page structure
The front-end structure of the fixture page was not structured.
Some rework was needed for showing all the matches in a single week.
Added cards to the page to show every match in a single card in a more
structured manner.This helps the user to understand and see the
matches and their data easier.