Implement Referee Assignment Process
The assignment page did not have the assignment feature working.
The referee list and the match were visible but nothing was happening
when clicked.Now when the TFF admin selects a referee from the list the
assignment is done by sending the update request to the database and
shows the newly assigned referee at the match page.
By following such an approach the user can easily see
his or her changes in the website instantly.