Update fixture page for reporter/retired referee
There was no designated fixture page for the reporters and retired
referees to see the matches.There was only a single fixture which was
created specially for the TFF admins.Now the reporters and retired
referees also have a fixture page which shows all the matches played
and the upcoming match week. By having a special page for
reporters/retired referees now the user can easily navigate
through the weeks for rating the referees of the played matches.