Create Responsive Assignment page for TFF admins
The referee assignment page was not responsive.

The admins were able to assign referee to matches but they could not see their changes in the frontend.

Now the admins can see their assignments instantly. When they press assign the name and the avatar of the assigned referees is updated.

By having such an approach now the admins can easily see their assignments instantly.