Add date and time data to fixture
The fixture page was not showing the date and time's of the respective
matches. Both admins and reporters/retired referees could see data
regarding the teams, referees and score but the date and time was
missing. Now the date and time information is also visible on the
fixture page for every match. By having such an approach the fixture
page frontend will be more structured and show the date and time data
which will be more understandable.