Add fixture to app bar
The app bar did not have the fixture tab.

User needed to go to the specific URL in order to reach the fixture.

Now with the added fixture tab the user can access the fixture page from the app bar.

By putting the fixture into the app bar the user will navigate faster between pages.