Change Reporter/Retired Referee app bar
While implementing the reporter fixture page the admin app bar was used
This caused a confusion for the user. Now the appbar is
changed to the specific reporter/retired referee app bar.
By doing that the reporters and retired referees
can only see the page indicators which they can enter