Add Reporter List Feature
This page allows TFF Admins to see a list of all reporters and access
their profile information. TFF Admins can also create a ban request for
any reporter from this page.The ban request feature is designed to
allow TFF Admins to quickly and easily ban reporters who are not
following the community guidelines. Overall, this commit aims to
improve the ability of TFF Admins to manage the community and ensure
that all reporters are following the guidelines.