Add ref create page route and update userControler
This route allows users to access the page where they can enter the
necessary information for creating a new referee. Additionally, the
query is updated on userController.js to properly handle the creation
of referees. This ensures that new referees are created accurately and
consistently. Overall, these changes make it easier for users to access
the referee creation page and create new referees, as well as ensure
that the application is able to handle this process properly.