Update to copyright component and referee list
In this commit, several updates have been made to the copyright
component and referee list on our website. The copyright component has
been updated to include website information, ensuring that our site's
credits are complete and up-to-date. The height of the referee list has
been decreased to improve the visual presentation of the page.
Additionally, styling has been added to enhance the overall appearance
of the page and ensure that everything is properly aligned and spaced