Add Backend for active referee
The backend for active referee creation and login has been updated to
allow for the creation of active referee accounts. These accounts are
generated with a random email and password using the OTP library.
Functions have been added to the backend to handle forcing a change of
OTP on the first login. After creation by an admin and changing the
OTP, users can use the account as usual.