Add page for recommendations system
In this commit, I have completed the page layout and design for the
recommendations system. The changes include a visually appealing design
and an easy-to-use layout. I have also integrated charts to provide
visual representations of the recommendations, and implemented a
recommendation algorithm to generate personalized recommendations for
users. The goal of these changes is to improve the user experience and
make it easier for users to interact with the recommendations system.
The integration of charts allows users to better understand and
visualize the recommendations, while the recommendation algorithm
ensures that users receive relevant recommendations.
I have completed the SQL queries for the recommendation
system. These queries will be used to retrieve and manipulate data as
needed to generate recommendations for users.I have also added the
queries to the backend of the application, where they will be used to
power the recommendation system. This will enable the recommendation
system to retrieve and manipulate data as needed to generate
personalized and relevant recommendations for users.