Add pages for resetting password
This commit adds the necessary components to enable users to reset
their password if they have forgotten it or need to update it for
security purposes. This feature should improve the user experience by
providing a way for users to reset their password if they have
forgotten it, rather than requiring them to contact an administrator or
create a new account. It also enhances the security of the system by
allowing users to update their password regularly