Add report form to match detail
There was no report field for the active referees.

The active referees were directed to the match detail page
however there was no report field. For them to enter their reports.

Now the active referees can select the match action from
a dropdown and enter their report regarding the selected action.

Their report is directly sent to the database.

By having such a field now the active referees
can have their comments on some of the actions of the match.