Add performance cards to referee rankings
I updated the ranking page design to include cards that visualize the
best performance of the week. These cards are intended to highlight top
performers and improve the visual appeal of the page.I added the cards
by modifying the front-end for the ranking page. I also made some small
adjustments to the layout and styling of the page to ensure that the
cards fit seamlessly into the design.