Add warning pages for rankings pages
This commit adds warning pages that are displayed when rankings are not
available or all matches have not been played for related week. These
warning pages provide information to the user about the current state
of the rankings, helping to improve the user experience by providing
more context and transparency when the rankings are not fully up to
date. By displaying these warning pages, we can better communicate to
the user why the rankings may not be complete and provide them with the
necessary information to understand the current state of the data.