Make UI developments for the profile page
The UI of the profile page was too crowded.

The user was able to see everything in the same row.
However, this was making the UI look too crowded.

Now the structure of the profile page is changed.
The page is implemented in a more vertical manner.

By making these improvements, The user can see all the information
but in a more understandable manner.