Add referee ranking button
There was a missing button for referee rankings in the fixture.

The admins were not able to see ranking button for the last week.

Now the referee ranking button is added to the page.

As a result the admins can click to the referee ranking button
in the last week.